feat: moved denormalization to omni_connector
